backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Pydantic
pydantic>=2.4.2
pydantic-settings>=2.0.3

# Database
sqlalchemy>=2.0.23

# Authentication (compatible versions for Python 3.13)
python-jose>=3.3.0
cryptography>=41.0.7
bcrypt>=4.1.2
passlib>=1.7.4

# File handling
aiofiles>=23.2.1

# PDF processing
PyPDF2>=3.0.1
pymupdf>=1.23.8

# Excel processing
openpyxl>=3.1.2
pandas>=2.1.3
numpy>=1.25.2

# Image processing
Pillow>=10.1.0

# AI/ML
google-generativeai>=0.3.2

# HTTP requests
requests>=2.31.0

# Redis (optional)
redis>=5.0.1

# Monitoring
prometheus-client>=0.19.0

# Logging
structlog>=23.2.0

# Development
python-decouple>=3.8